requirements.txt
pyrogram
tgcrypto
aiogram
apscheduler
mysql-connector-python
python-dotenv